+++ lib/vhcblade_widget.dart
@@ -1,0 +1,4 @@
+library vhcblade_widget;
+
+export 'src/widget/fade_through.dart';
+export 'src/widget/flexible_grid.dart';