+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 - Added VHCBladeTheme and default themes
 - Added VHCBladeThemeBloc
 
+## 0.1.1
+
+- Added Common Shared Widgets across apps
+    - FadeThroughWidgetSwitcher - creates a fade through transition when the child widget changes
+    - Flexible Grid - Similar to the GridView implementation, but children can be any size.
+
 ## 0.1.0
 
 - Initial Version